tools/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
requests==2.26.0
smmap==4.0.0
tqdm==4.62.2
urllib3==1.26.6
